fix(onboard): qualify policy receipt verification
