+++ src/lib/onboard/experimental/hermes-portable-policy-authority.ts
@@ -7,30 +7,11 @@ import { TextDecoder } from "node:util";
 import YAML from "yaml";
 
 import { parseOpenShellPolicy } from "../../policy/merge";
+import { isOpenShellNativeGpuBaselineEnrichment } from "../sandbox-gpu-route-policy";
 
 const UTF8 = new TextDecoder("utf-8", { fatal: true });
 const MAX_POLICY_BYTES = 256 * 1024;
 const NAME = /^[A-Za-z0-9][A-Za-z0-9._-]{0,255}$/u;
-const PROC_PATH = "/proc";
-const OPENSHELL_PROXY_REQUIRED_READ_ONLY_PATHS = new Set([
-  "/usr",
-  "/lib",
-  "/etc",
-  "/app",
-  "/var/log",
-  "/dev/urandom",
-]);
-const OPENSHELL_PROXY_REQUIRED_READ_WRITE_PATHS = new Set(["/tmp"]);
-const OPENSHELL_GPU_READ_ONLY_PATHS = new Set(["/run/nvidia-persistenced", "/usr/lib/wsl"]);
-const OPENSHELL_GPU_READ_WRITE_PATHS = new Set([
-  "/dev/nvidiactl",
-  "/dev/nvidia-uvm",
-  "/dev/nvidia-uvm-tools",
-  "/dev/nvidia-modeset",
-  "/dev/dxg",
-  PROC_PATH,
-]);
-const OPENSHELL_GPU_DEVICE_PATH = /^\/dev\/nvidia[0-9]+$/u;
 
 export interface HermesPortablePolicyCaptureResult {
   readonly status: number | null;
@@ -130,113 +111,6 @@ function semanticDigest(policy: Record<string, unknown>): string {
   return createHash("sha256")
     .update(JSON.stringify(canonical(policy)))
     .digest("hex");
-}
-
-function policyWithoutFilesystemPolicy(policy: Record<string, unknown>): Record<string, unknown> {
-  const result: Record<string, unknown> = Object.create(null);
-  for (const key of Object.keys(policy)) {
-    if (key !== "filesystem_policy") result[key] = policy[key];
-  }
-  return result;
-}
-
-function filesystemPolicyWithoutPaths(policy: Record<string, unknown>): Record<string, unknown> {
-  const result: Record<string, unknown> = Object.create(null);
-  for (const key of Object.keys(policy)) {
-    if (key !== "read_only" && key !== "read_write") result[key] = policy[key];
-  }
-  return result;
-}
-
-function readUniquePaths(
-  policy: Record<string, unknown>,
-  key: "read_only" | "read_write",
-): Set<string> | null {
-  const value = policy[key];
-  if (!Array.isArray(value) || value.some((entry) => typeof entry !== "string")) return null;
-  const paths = new Set(value);
-  return paths.size === value.length ? paths : null;
-}
-
-function isSubset(subset: ReadonlySet<string>, superset: ReadonlySet<string>): boolean {
-  return [...subset].every((entry) => superset.has(entry));
-}
-
-function setsOverlap(left: ReadonlySet<string>, right: ReadonlySet<string>): boolean {
-  return [...left].some((entry) => right.has(entry));
-}
-
-function isOpenShellNativeGpuBaselineEnrichment(
-  intended: Record<string, unknown>,
-  live: Record<string, unknown>,
-): boolean {
-  if (
-    semanticDigest(policyWithoutFilesystemPolicy(intended)) !==
-    semanticDigest(policyWithoutFilesystemPolicy(live))
-  ) {
-    return false;
-  }
-  const intendedFilesystem = intended.filesystem_policy;
-  const liveFilesystem = live.filesystem_policy;
-  if (
-    !intendedFilesystem ||
-    typeof intendedFilesystem !== "object" ||
-    Array.isArray(intendedFilesystem) ||
-    !liveFilesystem ||
-    typeof liveFilesystem !== "object" ||
-    Array.isArray(liveFilesystem)
-  ) {
-    return false;
-  }
-  const intendedFilesystemRecord = intendedFilesystem as Record<string, unknown>;
-  const liveFilesystemRecord = liveFilesystem as Record<string, unknown>;
-  if (
-    semanticDigest(filesystemPolicyWithoutPaths(intendedFilesystemRecord)) !==
-    semanticDigest(filesystemPolicyWithoutPaths(liveFilesystemRecord))
-  ) {
-    return false;
-  }
-  const intendedReadOnly = readUniquePaths(intendedFilesystemRecord, "read_only");
-  const intendedReadWrite = readUniquePaths(intendedFilesystemRecord, "read_write");
-  const liveReadOnly = readUniquePaths(liveFilesystemRecord, "read_only");
-  const liveReadWrite = readUniquePaths(liveFilesystemRecord, "read_write");
-  if (!intendedReadOnly || !intendedReadWrite || !liveReadOnly || !liveReadWrite) return false;
-  if (
-    setsOverlap(intendedReadOnly, intendedReadWrite) ||
-    setsOverlap(liveReadOnly, liveReadWrite)
-  ) {
-    return false;
-  }
-
-  // The native direct-GPU create policy deliberately omits /proc so OpenShell
-  // can add it read-write only after it observes the GPU devices. Treat that
-  // exact omission as the authority signal; ordinary/non-GPU policies retain
-  // /proc read-only and cannot use this exception.
-  if (intendedReadOnly.has(PROC_PATH) || intendedReadWrite.has(PROC_PATH)) return false;
-  if (
-    !isSubset(OPENSHELL_PROXY_REQUIRED_READ_ONLY_PATHS, intendedReadOnly) ||
-    !isSubset(OPENSHELL_PROXY_REQUIRED_READ_WRITE_PATHS, intendedReadWrite)
-  ) {
-    return false;
-  }
-  if (liveReadOnly.has(PROC_PATH) || !liveReadWrite.has(PROC_PATH)) return false;
-  if (!isSubset(intendedReadOnly, liveReadOnly) || !isSubset(intendedReadWrite, liveReadWrite)) {
-    return false;
-  }
-
-  for (const path of liveReadOnly) {
-    if (!intendedReadOnly.has(path) && !OPENSHELL_GPU_READ_ONLY_PATHS.has(path)) return false;
-  }
-  for (const path of liveReadWrite) {
-    if (
-      !intendedReadWrite.has(path) &&
-      !OPENSHELL_GPU_READ_WRITE_PATHS.has(path) &&
-      !OPENSHELL_GPU_DEVICE_PATH.test(path)
-    ) {
-      return false;
-    }
-  }
-  return true;
 }
 
 function rejectReservedCreateEntries(policy: Record<string, unknown>): void {
