+++ test/onboarding/onboard-extra-provider-reconciliation.test.ts
@@ -17,32 +17,35 @@ const onboardScriptMocksPath = JSON.stringify(
 );
 
 describe("onboard extra-provider reconciliation", () => {
-  it("attaches live user extras, prunes stale names, and converges registry state (#6501)", {
-    timeout: 90_000,
-  }, () => {
-    const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-extra-provider-reconcile-"));
-    try {
-      const fakeBin = path.join(tmpDir, "bin");
-      const scriptPath = path.join(tmpDir, "reconcile-provider-check.js");
-      const onboardPath = JSON.stringify(path.join(repoRoot, "src", "lib", "onboard.ts"));
-      const runnerPath = JSON.stringify(path.join(repoRoot, "src", "lib", "runner.ts"));
-      const registryPath = JSON.stringify(
-        path.join(repoRoot, "src", "lib", "state", "registry.ts"),
-      );
-      const preflightPath = JSON.stringify(
-        path.join(repoRoot, "src", "lib", "onboard", "preflight.ts"),
-      );
-      const credentialsPath = JSON.stringify(
-        path.join(repoRoot, "src", "lib", "credentials", "store.ts"),
-      );
-      const sandboxBaseImagePath = JSON.stringify(
-        path.join(repoRoot, "src", "lib", "sandbox-base-image.ts"),
-      );
+  it(
+    "attaches live user extras, prunes stale names, and converges registry state (#6501)",
+    {
+      timeout: 90_000,
+    },
+    () => {
+      const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-extra-provider-reconcile-"));
+      try {
+        const fakeBin = path.join(tmpDir, "bin");
+        const scriptPath = path.join(tmpDir, "reconcile-provider-check.js");
+        const onboardPath = JSON.stringify(path.join(repoRoot, "src", "lib", "onboard.ts"));
+        const runnerPath = JSON.stringify(path.join(repoRoot, "src", "lib", "runner.ts"));
+        const registryPath = JSON.stringify(
+          path.join(repoRoot, "src", "lib", "state", "registry.ts"),
+        );
+        const preflightPath = JSON.stringify(
+          path.join(repoRoot, "src", "lib", "onboard", "preflight.ts"),
+        );
+        const credentialsPath = JSON.stringify(
+          path.join(repoRoot, "src", "lib", "credentials", "store.ts"),
+        );
+        const sandboxBaseImagePath = JSON.stringify(
+          path.join(repoRoot, "src", "lib", "sandbox-base-image.ts"),
+        );
 
-      fs.mkdirSync(fakeBin, { recursive: true });
-      writeOkOpenshell(fakeBin, { readySandboxGet: true });
+        fs.mkdirSync(fakeBin, { recursive: true });
+        writeOkOpenshell(fakeBin, { readySandboxGet: true });
 
-      const script = String.raw`
+        const script = String.raw`
 const registry = require(${registryPath});
 const fixtureMocks = require(${onboardScriptMocksPath});
 registry.addExtraProvider("tavily-search");
@@ -162,83 +165,84 @@ const createReservedSandbox = () => createSandbox(
   process.exit(1);
 });
 `;
-      fs.writeFileSync(scriptPath, script);
+        fs.writeFileSync(scriptPath, script);
 
-      const result = spawnSync(process.execPath, [scriptPath], {
-        cwd: repoRoot,
-        encoding: "utf-8",
-        timeout: 30_000,
-        env: {
-          ...process.env,
-          HOME: tmpDir,
-          PATH: `${fakeBin}:${process.env.PATH || ""}`,
-          NEMOCLAW_NON_INTERACTIVE: "1",
-          NEMOCLAW_TEST_MANAGED_IMAGE_CATALOG: "1",
-          NEMOCLAW_SANDBOX_PREBUILD: "1",
-          OPENSHELL_GATEWAY_ENDPOINT: undefined,
-        },
-      });
+        const result = spawnSync(process.execPath, [scriptPath], {
+          cwd: repoRoot,
+          encoding: "utf-8",
+          timeout: 30_000,
+          env: {
+            ...process.env,
+            HOME: tmpDir,
+            PATH: `${fakeBin}:${process.env.PATH || ""}`,
+            NEMOCLAW_NON_INTERACTIVE: "1",
+            NEMOCLAW_TEST_MANAGED_IMAGE_CATALOG: "1",
+            NEMOCLAW_SANDBOX_PREBUILD: "1",
+            OPENSHELL_GATEWAY_ENDPOINT: undefined,
+          },
+        });
 
-      assert.equal(result.status, 0, result.stderr || result.error?.message);
-      const payloadLine = result.stdout
-        .trim()
-        .split("\n")
-        .reverse()
-        .find((line) => line.startsWith("{") && line.endsWith("}"));
-      assert.ok(payloadLine, `expected JSON payload in stdout:\n${result.stdout}`);
-      const payload = JSON.parse(payloadLine);
-      assert.deepEqual(payload.sandboxNames, ["my-assistant", "my-assistant"]);
+        assert.equal(result.status, 0, result.stderr || result.error?.message);
+        const payloadLine = result.stdout
+          .trim()
+          .split("\n")
+          .reverse()
+          .find((line) => line.startsWith("{") && line.endsWith("}"));
+        assert.ok(payloadLine, `expected JSON payload in stdout:\n${result.stdout}`);
+        const payload = JSON.parse(payloadLine);
+        assert.deepEqual(payload.sandboxNames, ["my-assistant", "my-assistant"]);
 
-      const createCommands = payload.commands.filter((entry: CommandEntry) =>
-        entry.command.includes("sandbox create"),
-      );
-      assert.equal(createCommands.length, 2, "expected one sandbox create command per attempt");
-      const [createCommand, retryCreateCommand] = createCommands;
-      assert.match(createCommand.command, /--provider brave-search/);
-      assert.match(createCommand.command, /--provider custom-provider/);
-      assert.match(createCommand.command, /--provider my-slack-bridge/);
-      assert.doesNotMatch(createCommand.command, /--provider tavily-search/);
-      assert.deepEqual(
-        createCommand.command.match(/--provider\s+\S+/g),
-        retryCreateCommand.command.match(/--provider\s+\S+/g),
-        "retry must preserve the exact filtered provider arguments",
-      );
+        const createCommands = payload.commands.filter((entry: CommandEntry) =>
+          entry.command.includes("sandbox create"),
+        );
+        assert.equal(createCommands.length, 2, "expected one sandbox create command per attempt");
+        const [createCommand, retryCreateCommand] = createCommands;
+        assert.match(createCommand.command, /--provider brave-search/);
+        assert.match(createCommand.command, /--provider custom-provider/);
+        assert.match(createCommand.command, /--provider my-slack-bridge/);
+        assert.doesNotMatch(createCommand.command, /--provider tavily-search/);
+        assert.deepEqual(
+          createCommand.command.match(/--provider\s+\S+/g),
+          retryCreateCommand.command.match(/--provider\s+\S+/g),
+          "retry must preserve the exact filtered provider arguments",
+        );
 
-      const providerProbes = payload.commands.filter((entry: CommandEntry) =>
-        entry.command.includes("provider get -g nemoclaw "),
-      );
-      assert.deepEqual(
-        providerProbes
-          .map((entry: CommandEntry) =>
-            entry.command.slice(entry.command.indexOf("provider get -g nemoclaw ")),
-          )
-          .sort(),
-        [
-          "provider get -g nemoclaw brave-search",
-          "provider get -g nemoclaw brave-search",
-          "provider get -g nemoclaw custom-provider",
-          "provider get -g nemoclaw custom-provider",
-          "provider get -g nemoclaw my-slack-bridge",
-          "provider get -g nemoclaw my-slack-bridge",
-          "provider get -g nemoclaw nvidia-prod",
-          "provider get -g nemoclaw nvidia-prod",
-          "provider get -g nemoclaw tavily-search",
-        ].sort(),
-      );
-      assert.equal(
-        payload.commands.some((entry: CommandEntry) =>
-          entry.command.includes("provider list -g nemoclaw --names"),
-        ),
-        false,
-        "provider-list snapshots must not control extra-provider attachment",
-      );
-      assert.deepEqual(payload.extraProviders, [
-        "brave-search",
-        "custom-provider",
-        "my-slack-bridge",
-      ]);
-    } finally {
-      fs.rmSync(tmpDir, { recursive: true, force: true });
-    }
-  });
+        const providerProbes = payload.commands.filter((entry: CommandEntry) =>
+          entry.command.includes("provider get -g nemoclaw "),
+        );
+        assert.deepEqual(
+          providerProbes
+            .map((entry: CommandEntry) =>
+              entry.command.slice(entry.command.indexOf("provider get -g nemoclaw ")),
+            )
+            .sort(),
+          [
+            "provider get -g nemoclaw brave-search",
+            "provider get -g nemoclaw brave-search",
+            "provider get -g nemoclaw custom-provider",
+            "provider get -g nemoclaw custom-provider",
+            "provider get -g nemoclaw my-slack-bridge",
+            "provider get -g nemoclaw my-slack-bridge",
+            "provider get -g nemoclaw nvidia-prod",
+            "provider get -g nemoclaw nvidia-prod",
+            "provider get -g nemoclaw tavily-search",
+          ].sort(),
+        );
+        assert.equal(
+          payload.commands.some((entry: CommandEntry) =>
+            entry.command.includes("provider list -g nemoclaw --names"),
+          ),
+          false,
+          "provider-list snapshots must not control extra-provider attachment",
+        );
+        assert.deepEqual(payload.extraProviders, [
+          "brave-search",
+          "custom-provider",
+          "my-slack-bridge",
+        ]);
+      } finally {
+        fs.rmSync(tmpDir, { recursive: true, force: true });
+      }
+    },
+  );
 });
